does not work sort wallets and delete

+++ src/main/java/com/test/bu/dao/WalletDao.java
@@ -17,7 +17,7 @@ public interface WalletDao {
 
    // void delete(int number);
 
-    void delete(int walletByNumber);
+    void delete(long walletByNumber);
 
     long getWalletsInDB();
