+++ src/main/java/com/test/bu/dao/WalletDaoIml.java
@@ -41,7 +41,7 @@ public class WalletDaoIml implements WalletDao {
     }
 
     @Override
-    public void delete(int walletByNumber) {
+    public void delete(long walletByNumber) {
 
         entityManager.remove(getWalletByNumber(walletByNumber));
     }
